useSmartSession hook
